@@ -6,18 +6,18 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
+import com.sz.charts.LineChart
 
-import com.androidplot.xy.XYPlot
 import smart_home.smarthome.entities.SensorData
 
 class EnvSensorsFragment(params: IGraphParameters) : SensorsFragment(R.layout.fragment_env_sensors, params, "env"), View.OnClickListener {
     private var mNextTempIntGraph: Button? = null
     private var mNextTempExtGraph: Button? = null
     private var mNextHumiGraph: Button? = null
-    private var mIntTempSensors: MutableList<List<IGraphData>>? = null
-    private var mExtTempSensors: MutableList<List<IGraphData>>? = null
-    private var mHumiSensors: MutableList<List<IGraphData>>? = null
-    private var mPresSensors: MutableList<IGraphData>? = null
+    private var mIntTempSensors: List<SensorData>? = null
+    private var mExtTempSensors: List<SensorData>? = null
+    private var mHumiSensors: List<SensorData>? = null
+    private var mPresSensor: SensorData? = null
     private var mIntTempSensorIdx: Int = 0
     private var mExtTempSensorIdx: Int = 0
     private var mHumiSensorIdx: Int = 0
@@ -25,11 +25,11 @@ class EnvSensorsFragment(params: IGraphParameters) : SensorsFragment(R.layout.fr
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View {
         val rootView = super.onCreateView(inflater, container, savedInstanceState)
-        mNextTempIntGraph = rootView!!.findViewById<Button>(R.id.next_temp_int_graph)
+        mNextTempIntGraph = rootView!!.findViewById(R.id.next_temp_int_graph)
         mNextTempIntGraph!!.setOnClickListener(this)
-        mNextTempExtGraph = rootView.findViewById<Button>(R.id.next_temp_ext_graph)
+        mNextTempExtGraph = rootView.findViewById(R.id.next_temp_ext_graph)
         mNextTempExtGraph!!.setOnClickListener(this)
-        mNextHumiGraph = rootView.findViewById<Button>(R.id.next_humi_graph)
+        mNextHumiGraph = rootView.findViewById(R.id.next_humi_graph)
         mNextHumiGraph!!.setOnClickListener(this)
         return rootView
     }
@@ -43,81 +43,40 @@ class EnvSensorsFragment(params: IGraphParameters) : SensorsFragment(R.layout.fr
     }
 
     private fun buildSensorLists(results: List<SensorData>) {
-        val intTempSensorsMap = mutableMapOf<String, MutableList<IGraphData>>()
-        val extTempSensorsMap = mutableMapOf<String, MutableList<IGraphData>>()
-        val humiSensorsMap = mutableMapOf<String, MutableList<IGraphData>>()
-        mPresSensors = mutableListOf()
-        results.forEach { data ->
-            if (data.data.containsKey("temp")) {
-                if (data.locationType != "ext") {
-                    val list = intTempSensorsMap[data.seriesColumnData]
-                    if (list == null) {
-                        intTempSensorsMap[data.seriesColumnData] = mutableListOf(data)
-                    } else {
-                        list.add(data)
-                    }
-                } else {
-                    val list = extTempSensorsMap[data.seriesColumnData]
-                    if (list == null) {
-                        extTempSensorsMap[data.seriesColumnData] = mutableListOf(data)
-                    } else {
-                        list.add(data)
-                    }
-                }
-            }
-            if (data.data.containsKey("humi")) {
-                val list = humiSensorsMap[data.seriesColumnData]
-                if (list == null) {
-                    humiSensorsMap[data.seriesColumnData] = mutableListOf(data)
-                } else {
-                    list.add(data)
-                }
-            }
-            if (data.data.containsKey("pres")) {
-                mPresSensors!!.add(data)
-            }
-        }
-
-        mIntTempSensors = ArrayList()
-        intTempSensorsMap.forEach { _, v ->
-            mIntTempSensors!!.add(v)
-        }
-        mExtTempSensors = ArrayList()
-        extTempSensorsMap.forEach { _, v ->
-            mExtTempSensors!!.add(v)
-        }
-        mHumiSensors = ArrayList()
-        humiSensorsMap.forEach { _, v ->
-            mHumiSensors!!.add(v)
-        }
+        mIntTempSensors = findSensors(results, { it != "ext" }, { it.contains("temp") })
+        mExtTempSensors = findSensors(results, { it == "ext" }, { it.contains("temp") })
+        mHumiSensors = findSensors(results, { true }, { it.contains("humi") })
+        mPresSensor = findSensors(results, { true }, { it.contains("pres") }).firstOrNull()
     }
 
     private fun showResults(v: View) {
         showIntTempSensors(v)
         showExtTempSensors(v)
         showHumiSensors(v)
-        val plot = v.findViewById<XYPlot>(R.id.plot_pres)
-        Graph.buildGraph(plot, mPresSensors!!, "pres", false, !params!!.getData().mUsePeriod, 2, 2)
+        val plot = v.findViewById<LineChart>(R.id.plot_pres)
+        Graph.buildGraph(plot, mPresSensor, "pres", false, !params!!.getData().mUsePeriod, false)
     }
 
     private fun showIntTempSensors(v: View) {
         if (mIntTempSensors != null && mIntTempSensorIdx < mIntTempSensors!!.size) {
-            val plot = v.findViewById<XYPlot>(R.id.plot_temp_int)
-            Graph.buildGraph(plot, mIntTempSensors!![mIntTempSensorIdx], "temp", false, !params!!.getData().mUsePeriod,2, 2)
+            val plot = v.findViewById<LineChart>(R.id.plot_temp_int)
+            Graph.buildGraph(plot, mIntTempSensors!![mIntTempSensorIdx], "temp", false, !params!!.getData().mUsePeriod, false)
         }
     }
 
     private fun showExtTempSensors(v: View) {
+        val plot = v.findViewById<LineChart>(R.id.plot_temp_ext)
         if (mExtTempSensors != null && mExtTempSensorIdx < mExtTempSensors!!.size) {
-            val plot = v.findViewById<XYPlot>(R.id.plot_temp_ext)
-            Graph.buildGraph(plot, mExtTempSensors!![mExtTempSensorIdx], "temp", false, !params!!.getData().mUsePeriod, 2, 2)
+            Graph.buildGraph(plot, mExtTempSensors!![mExtTempSensorIdx], "temp", false, !params!!.getData().mUsePeriod, false)
+        } else {
+            plot.clearSeries()
         }
     }
 
     private fun showHumiSensors(v: View) {
         if (mHumiSensors != null && mHumiSensorIdx < mHumiSensors!!.size) {
-            val plot = v.findViewById<XYPlot>(R.id.plot_humi)
-            Graph.buildGraph(plot, mHumiSensors!![mHumiSensorIdx], "humi", false, !params!!.getData().mUsePeriod, 2, 2)
+            val plot = v.findViewById<LineChart>(R.id.plot_humi)
+            Graph.buildGraph(plot, mHumiSensors!![mHumiSensorIdx], "humi", false, !params!!.getData().mUsePeriod, false)
         }
     }
 
@@ -146,7 +105,7 @@ class EnvSensorsFragment(params: IGraphParameters) : SensorsFragment(R.layout.fr
     }
 
     private fun showNextTempIntGraph() {
-        if (mIntTempSensors != null && mIntTempSensors!!.size > 0) {
+        if (mIntTempSensors != null && mIntTempSensors!!.isNotEmpty()) {
             mIntTempSensorIdx++
             if (mIntTempSensorIdx >= mIntTempSensors!!.size) {
                 mIntTempSensorIdx = 0
@@ -156,7 +115,7 @@ class EnvSensorsFragment(params: IGraphParameters) : SensorsFragment(R.layout.fr
     }
 
     private fun showNextTempExtGraph() {
-        if (mExtTempSensors != null && mExtTempSensors!!.size > 0) {
+        if (mExtTempSensors != null && mExtTempSensors!!.isNotEmpty()) {
             mExtTempSensorIdx++
             if (mExtTempSensorIdx >= mExtTempSensors!!.size) {
                 mExtTempSensorIdx = 0
@@ -166,7 +125,7 @@ class EnvSensorsFragment(params: IGraphParameters) : SensorsFragment(R.layout.fr
     }
 
     private fun showNextHumiGraph() {
-        if (mHumiSensors != null && mHumiSensors!!.size > 0) {
+        if (mHumiSensors != null && mHumiSensors!!.isNotEmpty()) {
             mHumiSensorIdx++
             if (mHumiSensorIdx >= mHumiSensors!!.size) {
                 mHumiSensorIdx = 0
