@@ -2,7 +2,6 @@ package smart_home.smarthome
 
 import android.app.Activity
 import android.app.DatePickerDialog
-import android.content.Intent
 import android.os.Bundle
 import android.os.Parcel
 import android.os.Parcelable
@@ -12,11 +11,12 @@ import android.widget.AdapterView
 import android.widget.Button
 import android.widget.Spinner
 import android.widget.TextView
+import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 
-class FiltersActivity() : AppCompatActivity(), View.OnClickListener, AdapterView.OnItemSelectedListener {
+class FiltersActivity : AppCompatActivity(), View.OnClickListener, AdapterView.OnItemSelectedListener {
     companion object {
         val UI_DATE_FORMAT: DateTimeFormatter = DateTimeFormatter.ofPattern("dd MMM yyyy")
     }
@@ -74,20 +74,20 @@ class FiltersActivity() : AppCompatActivity(), View.OnClickListener, AdapterView
         setContentView(R.layout.filters_activity)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
-        val select_date_start = findViewById<Button>(R.id.select_date_start)
-        val select_date_end = findViewById<Button>(R.id.select_date_end)
+        val selectDateStart = findViewById<Button>(R.id.select_date_start)
+        val selectDateEnd = findViewById<Button>(R.id.select_date_end)
         val set = findViewById<Button>(R.id.set)
-        val filter_type = findViewById<Spinner>(R.id.filter_type)
+        val filterType = findViewById<Spinner>(R.id.filter_type)
         val period = findViewById<Spinner>(R.id.period)
 
-        select_date_start.setOnClickListener(this)
-        select_date_end.setOnClickListener(this)
+        selectDateStart.setOnClickListener(this)
+        selectDateEnd.setOnClickListener(this)
         set.setOnClickListener(this)
-        filter_type.onItemSelectedListener = this
+        filterType.onItemSelectedListener = this
 
         mData = intent.getParcelableExtra("data")!!
 
-        filter_type.setSelection(if (mData.mUsePeriod) {0} else {1})
+        filterType.setSelection(if (mData.mUsePeriod) {0} else {1})
 
         val periods = resources.getIntArray(R.array.periodValues)
         var idx = 0
@@ -99,14 +99,21 @@ class FiltersActivity() : AppCompatActivity(), View.OnClickListener, AdapterView
             idx++
         }
 
+        onBackPressedDispatcher.addCallback(this, object : OnBackPressedCallback(true) {
+            override fun handleOnBackPressed() {
+                setResult(mData.mResult, intent)
+                finish()
+            }
+        })
+
         updateDate()
     }
 
     private fun updateDate() {
-        val date_start = findViewById<TextView>(R.id.date_start)
-        val date_end = findViewById<TextView>(R.id.date_end)
-        date_start.text = mData.mDateStart.format(UI_DATE_FORMAT)
-        date_end.text = mData.mDateEnd.format(UI_DATE_FORMAT)
+        val dateStart = findViewById<TextView>(R.id.date_start)
+        val dateEnd = findViewById<TextView>(R.id.date_end)
+        dateStart.text = mData.mDateStart.format(UI_DATE_FORMAT)
+        dateEnd.text = mData.mDateEnd.format(UI_DATE_FORMAT)
     }
 
     public override fun onResume() {
@@ -114,15 +121,11 @@ class FiltersActivity() : AppCompatActivity(), View.OnClickListener, AdapterView
         mData.mResult = Activity.RESULT_CANCELED
     }
 
-    override fun onBackPressed() {
-        setResult(mData.mResult)
-        finish()
-    }
-
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when (item.itemId) {
             android.R.id.home -> {
-                onBackPressed()
+                setResult(mData.mResult, intent)
+                finish()
                 return true
             }
         }
@@ -130,68 +133,71 @@ class FiltersActivity() : AppCompatActivity(), View.OnClickListener, AdapterView
     }
 
     override fun onClick(v: View?) {
-        val select_date_start = findViewById<Button>(R.id.select_date_start)
-        val select_date_end = findViewById<Button>(R.id.select_date_end)
+        val selectDateStart = findViewById<Button>(R.id.select_date_start)
+        val selectDateEnd = findViewById<Button>(R.id.select_date_end)
 
-        if (v == select_date_start) {
-            val dialog = DatePickerDialog(this, { _, year, monthOfYear, dayOfMonth ->
-                mData.mDateStart = LocalDate.of(year, monthOfYear + 1, dayOfMonth)
-                if (mData.mDateEnd.isBefore(mData.mDateStart)) {
-                    mData.mDateEnd = mData.mDateStart
-                }
-                updateDate()
-            }, mData.mDateStart.year, mData.mDateStart.monthValue - 1, mData.mDateStart.dayOfMonth)
-            dialog.show()
-        } else if (v == select_date_end) {
-            val dialog = DatePickerDialog(this, { _, year, monthOfYear, dayOfMonth ->
-                mData.mDateEnd = LocalDate.of(year, monthOfYear + 1, dayOfMonth)
-                if (mData.mDateStart.isAfter(mData.mDateEnd)) {
-                    mData.mDateStart = mData.mDateEnd
-                }
-                updateDate()
-            }, mData.mDateStart.year, mData.mDateStart.monthValue - 1, mData.mDateStart.dayOfMonth)
-            dialog.show()
-        } else { // Generate button
-            val filter_type = findViewById<Spinner>(R.id.filter_type)
-            val filterType =  resources.getStringArray(R.array.filterTypeValues)[filter_type.selectedItemPosition]
-            mData.mUsePeriod = filterType == "Period"
-            val period = findViewById<Spinner>(R.id.period)
-            mData.mPeriod = resources.getIntArray(R.array.periodValues)[period.selectedItemPosition]
-            val data: Intent = Intent()
-            data.putExtra("data", mData)
-            setResult(Activity.RESULT_OK, data)
-            finish()
+        when (v) {
+            selectDateStart -> {
+                val dialog = DatePickerDialog(this, { _, year, monthOfYear, dayOfMonth ->
+                    mData.mDateStart = LocalDate.of(year, monthOfYear + 1, dayOfMonth)
+                    if (mData.mDateEnd.isBefore(mData.mDateStart)) {
+                        mData.mDateEnd = mData.mDateStart
+                    }
+                    updateDate()
+                }, mData.mDateStart.year, mData.mDateStart.monthValue - 1, mData.mDateStart.dayOfMonth)
+                dialog.show()
+            }
+            selectDateEnd -> {
+                val dialog = DatePickerDialog(this, { _, year, monthOfYear, dayOfMonth ->
+                    mData.mDateEnd = LocalDate.of(year, monthOfYear + 1, dayOfMonth)
+                    if (mData.mDateStart.isAfter(mData.mDateEnd)) {
+                        mData.mDateStart = mData.mDateEnd
+                    }
+                    updateDate()
+                }, mData.mDateStart.year, mData.mDateStart.monthValue - 1, mData.mDateStart.dayOfMonth)
+                dialog.show()
+            }
+            else -> { // Generate button
+                val filterTypeView = findViewById<Spinner>(R.id.filter_type)
+                val filterType =  resources.getStringArray(R.array.filterTypeValues)[filterTypeView.selectedItemPosition]
+                mData.mUsePeriod = filterType == "Period"
+                val period = findViewById<Spinner>(R.id.period)
+                mData.mPeriod = resources.getIntArray(R.array.periodValues)[period.selectedItemPosition]
+                intent.putExtra("data", mData)
+                setResult(Activity.RESULT_OK, intent)
+                finish()
+            }
         }
     }
 
     override fun onItemSelected(parent: AdapterView<*>?, view: View?, position: Int, id: Long) {
-        val filter_type = findViewById<Spinner>(R.id.filter_type)
+        val filterTypeView = findViewById<Spinner>(R.id.filter_type)
         val periodLabel = findViewById<TextView>(R.id.period_label)
         val period = findViewById<Spinner>(R.id.period)
-        val select_date_start = findViewById<Button>(R.id.select_date_start)
-        val select_date_end = findViewById<Button>(R.id.select_date_end)
-        val date_start = findViewById<TextView>(R.id.date_start)
-        val date_end = findViewById<TextView>(R.id.date_end)
-        val filterType =  resources.getStringArray(R.array.filterTypeValues)[filter_type.selectedItemPosition]
+        val selectDateStart = findViewById<Button>(R.id.select_date_start)
+        val selectDateEnd = findViewById<Button>(R.id.select_date_end)
+        val dateStart = findViewById<TextView>(R.id.date_start)
+        val dateEnd = findViewById<TextView>(R.id.date_end)
+        val filterType =  resources.getStringArray(R.array.filterTypeValues)[filterTypeView.selectedItemPosition]
         if (filterType == "Period") {
             periodLabel.isEnabled = true
             period.isEnabled = true
-            date_start.isEnabled = false
-            select_date_start.isEnabled = false
-            date_end.isEnabled = false
-            select_date_end.isEnabled = false
+            dateStart.isEnabled = false
+            selectDateStart.isEnabled = false
+            dateEnd.isEnabled = false
+            selectDateEnd.isEnabled = false
         } else {
             periodLabel.isEnabled = false
             period.isEnabled = false
-            date_start.isEnabled = true
-            select_date_start.isEnabled = true
-            date_end.isEnabled = true
-            select_date_end.isEnabled = true
+            dateStart.isEnabled = true
+            selectDateStart.isEnabled = true
+            dateEnd.isEnabled = true
+            selectDateEnd.isEnabled = true
         }
     }
 
     override fun onNothingSelected(parent: AdapterView<*>?) {
-        val filter_type = findViewById<Spinner>(R.id.filter_type)
-        filter_type.setSelection(0)
+        val filterType = findViewById<Spinner>(R.id.filter_type)
+        filterType.setSelection(0)
     }
 }
